Change Symbol fix: use MacroCommand
