+++ Editor/Gui/Graph/GraphCanvas.cs
@@ -835,20 +835,7 @@ namespace T3.Editor.Gui.Graph
                 var position = selectedChildUis.Count == 1 ? selectedChildUis[0].PosOnCanvas : InverseTransformPositionFloat(ImGui.GetMousePos());
                 SymbolBrowser.OpenAt(position, null, null, false, startingSearchString, symbol =>
                     {
-                        var nextSelection = new List<SymbolChild>();
-                        foreach (var sel in selectedChildUis)
-                        {
-                            var result = ChangeSymbol.ChangeOperatorSymbol(sel, symbol);
-                            if (result != null)
-                                nextSelection.Add(result);
-                        }
-                        NodeSelection.Clear();
-                        nextSelection.ForEach(symbolChild => {
-                                var childUi = SymbolUiRegistry.Entries[symbolChild.Parent.Id].ChildUis.SingleOrDefault(ui => ui.SymbolChild == symbolChild); // need better map?
-                                var instance = CompositionOp.Children.Single(c2 => c2.SymbolChildId == symbolChild.Id);
-                                if(childUi != null && instance != null)
-                                    NodeSelection.AddSymbolChildToSelection(childUi, instance); 
-                            });
+                        ChangeSymbol.ChangeOperatorSymbol(CompositionOp, selectedChildUis, symbol);
                     });
             }
